FEN translation, info to the side, etc
FEN translation is ready (not usable since it is not implemented), there is now some info to the side of the board, like the pieces that have been eaten and the moves and half-moves passed. Getters and setters introduced. Castling is no longer allowed while checked. Printing the board got a little bit complex, but that's just how it is.